% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPASS-package.R
\docType{package}
\name{COMPASS-package}
\alias{COMPASS-package}
\title{COMPASS (Combinatorial Polyfunctionality Analysis of Single-Cells)}
\description{
This package implements a model for the analysis of polyfunctionality
in single-cell cytometry experiments. The model effectively
identifies combinations of markers that are differentially expressed
between samples of cells subjected to different stimulations.
}
\seealso{
\itemize{
  \item \code{\link{COMPASSContainer}}, for information on getting your
  cytometry data into a suitable format for use with \code{COMPASS},
  \item \code{\link{COMPASS}}, for the main model fitting routine.
  }
}
\author{
\strong{Maintainer}: Greg Finak \email{gfinak@fhcrc.org}

Authors:
\itemize{
  \item Lynn Lin \email{llin@fhcrc.org}
  \item Kevin Ushey \email{kushey@fhcrc.org}
}

Other contributors:
\itemize{
  \item Raivo Kolde \email{rkolde@gmail.com} (Author of 'pheatmap', which was modified and now used internally in COMPASS) [contributor]
}

}
